feat: Add support for SSH keys in setup script
The code changes add support for SSH keys in the setup script. It checks if the private and public SSH keys already exist or are provided via environment variables. If they don't exist, it generates an ED25519 SSH key pair and displays the public key for the user to copy. The user is prompted to add the SSH key to their GitHub account.